Extract DorSolrDocumentBuilder from DorHarvester; fixes #77
